Add files via upload
My basic project is GPA calculator but a little bit developed version. There will 
be ready class names and their key words, if we enter key word for class, (and also 
we will be able to define the number of classes for a semester) program will ask 
from us grades of relevant classes. Finally, program will give us GPA for a 
semester.